NI defs use low projections; restructure proofs
The main point of this change is to both to make the security definition
provable (while still ensuring security) and also to restructure the proofs.
The main change is to:
- move away from a statement of the theorem that uses a low-equivalence
relation directly and instead use one based on low projections.
- write a set of lemmas about taking steps to/from the low projections of
states, and use these to prove the single step theorem.
- the system-wide step function is also extended with a system-wide NOP

As side-effects:

These changes are fairly sweeping, so many of the old theorems are invalidated
The definitions of step functions in EvAugSemantics are rewritten to be
clearer/cleaner/better-separated:
- The single-node function is now (nid -> call -> state -> state -> event)
rather than (nid -> call -> trace -> trace) which makes it closer to the step
function that does not emit events
- A separate relation builds traces using the individual steps on states.

The main idea behind this change:

The heart of the problem with the previous definition and proof structure
is in the theorem about a single step of noninterference for an individual node.
An unusual aspect of this proof compared to one for a programming language is
that for a language, the single step theorem would compare two executions
(each with different values for secrets) for a single particular command that
is used in both executions, but in this case the command executed by a node
might be a secret itself. The fact that the command is a secret shows up in
the old low-equivalence definition: when the label of the node does not flow
to the observer, it relates any two nodes with the same label (but possibly
differing commands). This presents a problem in the old single-step theorem
(possibilistic_ni_1step_node) which we were attempting to do by cases on the
command in the first execution, and then by cases on whether the label of the
node that was executed in the first exectution flowsTo the observer or not.
In the flowsTo case, the node referred to by the chosen node id is the equal
in both executions. But in the not-flows-to case, we know nothing about what
the node looks like in the other execution (its contents are a secret).
So the command executed by the node could be anything, even an invalid
command that cannot take a step. So this old theorem with the old security
definitions is not true.

This PR attempts to fix this problem by using low projections in the
definition of security and a new proof structure. The low projections are
functions of type: {A: Type} -> label -> A -> A, that removes all the secrets
(above the label) from the thing of type A. Examples of things are nodes,
channels, and events. The new proof structure no longer has the
possibilistic_ni_1step_node theorem, but instead uses a new set of theorems
about steps to/from low projections over states (step_implies_lowproj_steps_leq,
low_proj_steps_implies_leq_step) in the proof for possibilistic_ni_1step
(which is about a system-wide step rather than a single-node step).
The (new) proof for the lemma possibilistic_ni_1step is complete
(aside from necessitating some obviously true lemmas about transitivity of
low-equivalences).

An intuition as to how this (should) fix rather than just move the problem is
that now the potential issue involving unobservable steps might show up in
the low_proj_steps_implies_leq_step lemma. However, when the step is a
command run by a secret node, the low-projection will: (1) only modify secret
state, and (2) only emit NOP-events (the low projection of a secret event is a
NOP). So for execution 2, we can choose the state and event that results from
a system-wide NOP because this should have equivalent behavior as the secret
step.