Nflows case for low_proj_steps_implies_leq_step
This includes a proof of low_proj_steps_implies_leq_step for the case in
which a NodeStep is taken, and the particular node executing the step has a
label that is not visible to the observer. This was the case that seemed most
likely to determine whether or not the proof strategy will work
out, so the proof suggests it will.

Some new admitted theorems are introduced.

Some cleanup related to the placement of theorems was done.