Lowproj step WriteChannel FlowsTo case (#1535)
This commit completes the proof of the lemma low_proj_steps_implies_leq_step for just the WriteChannel call and just in the case where the label of the node making the call flows to the label of the observer. It is expected that for the other calls, the proofs might look similar. (However, the read call may be an interesting case because perhaps we need to do something special to show that read handles are linear).

Some new smaller theorems without proofs yet are introduced.

Some other smaller theorems are also either finished or progress is made on these proofs as well.

The constructor names for step_node_ev are disambiguated from those of step_node.