Finish low_proj_steps_implies... + ..._unwind_t; tactics (#1578)
This commit finishes the proofs of low_proj_steps_implies_leq_step and possibilistic_ni_unwind_t. 

A substantial part of the proofs for both of the theorems about stepping into / out of low-projected states involved:
- doing some calculation to figure out the final state
- econstructor; eauto - ing to apply the constructors for the various kinds of steps
- calling the right "unwind" functions
- infrequently, a few other interesting bits.

We had a good hint database for unwind and this updates that.

There is now an additional automated tactic for doing the constructor calls, which is folded into crush.

The proofs are relatively close to just mentioning the calculation (which is done manually), writing the rare interesting bits, and using the automated tactics. The need to manually unfold variables that are defined by let statements in the spec and one place where the unwind hint database seem like the last few unfortunate bits.

(contributions from @jadephilipoom here as well)