Implement type safe entrypoint method
This modifies the `entrypoint` macro to specify a static type for the
incoming messages, and builds a type safe `Receiver` from the underlying
raw channel.

It also allowed removing the `app_config_map` method, since now any
method that wants to obtain a config map, may just specify this as part
of the type of incoming messages, and read the first message from it.

Note in some cases there is still a bit of stuttering, where the type
inference is not strong enough to figure out the full type of the
`Receiver`, but I think it may be improved in the future.