Partial proof of unobservable step lemma
- but the eautos found an inconsistent proof:
https://github.com/coq/coq/issues/8215
It turns out with the
present definitions, the lemma is not true. The definitions of
node_state_low_eq and chan_state_low_eq need to be changed so that None
is mapped to the low projection. Doing so would make it impossible for a
public observer to tell whether or not a secret node or channel was
created or destroyed. This is needed to make the unobservable step lemma
provable because otherwise, channel create turns an unoccouped handle
into one that is occupied. The case is similar with node create.

- The call for ReadChannel needs to be changed so that label_of(ch) =
label_of(node) rather than label_of(ch) << label_of(node) because
channel reads are destructive. See also:
https://github.com/project-oak/oak/issues/1197#issuecomment-712180898
(which is still being discussed)