Fix state projection defs; finish unobs_step lemma
The definition of low projection on the node and channel part of the
state has been amended so that handles and node IDs that are unoccupied
(i.e. map to None) in the real state are instead mapped to the empty
channel/node in the projection. This means low-adversaries cannot
distinguish between unoccupied slots and secret objects, which means
creating a new node/channel cannot be observed.

This also makes the unobservable_step lemma provable.

A lemma in NIUtils (proj_node_state_to_proj_n) is no longer true with
this change in the definitions. So we need to figure out a new way to finish
the proof that uses it.