Implement Decodable and Encodable for Either type
The `Either` type comes from the https://docs.rs/either/1.6.1/either/
crate instead of being re-implemented from scratch, as that has useful
methods already defined on it, and it is pretty lightweight.

This may be used as foundation for more interesting patterns related to
multiplexing several message types on the same channel, e.g. init and
command messages.