Amend definitions of labeled things. (#1642)
The main reason for this change is to make the proj_node_state_to_proj_n
provable, while still supporting creating/destroying nodes and channels
that are labeled secret.

The main change is that for objects that have labels:
- labels are now decoupled from the rest of the object
- the objects are replaced by options, and empty versions are really
none. This is done with the definition "labeled" in RuntimeModel.v

To give confidence that this change makes forward progress, note that
- secret labeled objects holding a None are indistinguishable from real
ones, so this does not undo the change in
[#1603](https://github.com/project-oak/oak/pull/1603)
- If a nid/handle points to a real object in a projected state, it does
also guarantee the same is true in the non-projected state, making the
troublesome theorem provable.

As a side-effect the definitions of several theorems in NIUtilTheorems
get better reuse, and the low projections are also probably more
trustworthy since they are more clearly distinguished from real objects.

This change does unfortunately break a lot of the proofs.