Allow creating node entrypoint with init data
Introduce a `WithInit` trait implemented by structs that correspond to
nodes that require receiving initialization data over their inbound
channel, before accepting command from the wrapped channel.

Currently there are two entrypoint-related macros, one for structs that
impl `Default` and do not require init data, and the new one for structs
that impl `WithInit`, which expect an `InitWrapper` message. It may be
possible to unify them in the future with some more macro magic, or we
may just decide to make everything impl `WithInit`, using `()` as init
in case the node does not require init data, but for now I think the
separation makes sense.

Ref #1639