Prove step_implies_lowproj_steps_leq (#1672)
Notably, the way channel append works when writing to a channel needed
to be changed in order to make this provable. Because it is possible for
a public node to send a message to a secret channel, and because a
public observer cannot distinguish between a handle that points to a
secret channel and a handle that points to nothing (becuase the channel
was deleted for example), the write call cannot check whether or not the
handle is allocated because this would cause a leak. This potential leak made
step_implies_lowproj unprovable without changing the semantics.

The change is in chan_append_labeled and the semantics of the write call.
If the write call is to a handle that points to nothing, the write has no
effect and no error is thrown. See chan_append_labeled.

For similar reasons, the checks for "fresh" node ids and handles had to
be removed, though I think these don't correspond to anything in the
real implementation anyway.