Rewrite state_low_projection to obviate functional extensionality.
Rewrite state_low_projection to obviate functional extensionality.

See also discussion in [PR 1685](https://github.com/project-oak/oak/pull/1685) where @jadephilipoom suggested this change.

Like any axiom, functional extensionality can be dangerous to add to a proof, so it is useful to avoid adding it if possible. This change avoids that by writing the low-equivalence relation on the state (and the parts of the state that are functions) so that they are low-equivalent if applying these functions to any argument gives low-equivalent results, rather than defining them in terms of (leibniz) equality directly. 

The proof state is also no worse than it was before.

Co-authored-by: Jade Philipoom <jadep@google.com>