Generate simpler gRPC methods
Instead of going back and forth between the generated code and the code
in `oak::grpc`, this change simplifies the logic of the dispatcher by
returning pointers to the appropriate methods based on method name. This
also allows accessing and manipulating those methods from other code,
enabling also reflection-like operations based on method names.

I think the types and lifetimes may be simplified even more, and also a
similar refactoring may be done for client methods, but I'll leave those
for a separate PR.

Ref #1691.