Use protobuf for machine_learning example
This is the only example that uses `impl oak::grpc::ServerNode`
directly, which I am going to refactor / delete as part of #1691, so
this makes it a blocker.

I think we just implemented this manually in the early days before we
had protobuf support.

I naively made all the methods return an empty request, which is in line
with the previous implementation, but not very nice.

I think we should probably delete this example anyways, so I'm just
doing the minimal amount of work for this refactoring.