Fix leak in semantics; Make labels partially secret(#1699)
There are two main changes here:
1) a change to the semantics of the runtime to avoid a leak
2) a change to the formulation of low projections that makes labels themselves
partially secret in order to make NI provable while still supporting dynamic
node/channel creation

(1) The semantic change:
Notably, the way channel append works when writing to a channel needed
to be changed in order to make this provable. Because it is possible for
a public node to send a message to a secret channel, and because a
public observer cannot distinguish between a handle that points to a
secret channel and a handle that points to nothing (becuase the channel
was deleted for example), the write call cannot check whether or not the
handle is allocated because this would cause a leak. This potential leak made
step_implies_lowproj unprovable without changing the semantics.

The change is in chan_append_labeled and the semantics of the write call.
If the write call is to a handle that points to nothing, the write has no
effect and no error is thrown. See chan_append_labeled.


(2) The formalization change:
The definition of low projection is changed so that labels are *not* public,
but instead they are partially secret. An observer at level ell can see:
    - whether the label of an object flows to ell, for any object.
    - the precise label of the object if its label does flow to ell.
Putting it another way, the label ell' of a secret object is not observable,
but whether or not (ell' <<L ell) is observable.

This is necessary to support dynamic node/channel creation. See comment in
LowEquivalences.v for more detail.