runtime: fix unit tests
Currently, only Nodes that have a public_untrusted label are allowed to create
channels or other Nodes, so ensure that any test that exercises node/channel
creation involves a test node with public_untrusted label.

Also, it's not possible to create a logging pseudo-Node with a label other than
public_untrusted, so use a gRPC pseudo-Node instead (which in turn requires
the test to include a gRPC configuration).