Split low projection definitions + model
The two different definitions of low projection and two different models
discussed in (issue #1722)[https://github.com/project-oak/oak/issues/1722]
are both alive in the code up to some smaller theorems.

The code redundancy is not ideal at the moment, so if these are really
meant to stay alive forever, then that should be improved at some point.