Finish NIUtils.v proofs (#1777)
The definition of fresh_han and fresh_chan needed to be changed so
that the labels of unused objects is public. For the other definition
where labels are partially secret, I think this would need to stay
"top". Either way, fresh_han and fresh_chan are not really parts of the
implementation; in the implementation I think these objects are just
genuinely newly created objects.

There is also some leftover fixup that didn't quite make it into the leftover PR.