Downgrading Security Specification (#1844)
This is a candidate specification of a security condition that does permit downgrades and ensures that the only way around the information flow labels is through these downgrades. There is just a partial specification of the model that supports downgrades that includes just one call so that the proof effort can focus on the proof structure and the model can be scaled up later.