Prove partially-secret PNI up through unwind (#1871)
Partially-secret means a node can only see labels that flow to its
label (i.e. are at most as secret as the node is), so labels are
partially secret. This commit makes progress towards proving
possibilistic non-interference with a partially-secret runtime model.

Some of the added content might be refactorable if we decide to
maintain both versions of Possibilistic NI in the long term.