Solving finite (non-semi!)-lattices in HORNSAT
Using the ADT representation from the last commit we would have a way to
encode (and solve the needed relations) for join semilattices. If we
wanted to have a richer lattice structure like FLAM or the DC label
model, semi-lattices would not be sufficient. In other words, we can't
have both meets and joins with that other representation (that would be
reducible to SAT, and in Datalog or even Prolog, only HORNSAT is
solvable).

This commit tries to pickup the hypercube label model from HyperFlow to
solve that problem. Here lattices are not modeled
purely relationally: lattice elements have an inner structure which is a
point within an n-dimensional space. Elements e1, e2 are ordered if each
coordinate of e1 is (numerically) less than the corresponding coordinate
of e2. Meets and joins are the dimension-wise min/max over the
coordinates. Translations from FLAM to this representation are also given in
the paper.

While this does give us a way to represent richer label models in
HORNSAT, evaluation seems relatively slow even for individual comparisons because
datalog uses bottom-up evaluation. Ostensibly this is one thing that
would be faster in prolog.