Unify wasm compilation flags
This removes some of the logic around compiling Wasm modules to a
specific out dir, but I think it makes more sense to keep the default
location and modify the app manifest files to point to it rather than
convincing cargo to put compiled artifacts in non-standard folders, even
if the manifests end up being slightly longer.

- Fix the path of some of the internal dependencies, since some tests
started randomly failing, presumably because of some interference with
externally published crates with the same names (namely `prost`).