Key / value lookup example and generic client (#2095)
The intention is to keep this example around as a building block for
other discussions, and also have something that does not necessarily
need to evolve over time based on other requirements.

The generic client may be used for demos and debugging instances of Oak
Functions running either remotely or locally.

Fix #2093
Ref #2094