*** Enter description below (required): Prove most of PNI with partially-secret labels
Completes all of Possibilistic non-interference for partially-secret
labels with the exception of one critical theorem, which relies on a
fact that is no longer true with partially secret labels. The erroneous
fact is commented out, and the critical theorem is left Admitted.

PNI should still be provable, but more work will be required to find a
workaround for that fact and complete the proof.