Run the Rust-related CI steps only for the crates affected by the changes in the PR (#2469)
Adds a top-level `scope` flag to all commands. It has three variants:
`--scope=all`, `--scope=diff_to_main`, `--scope=commits:<N>`. The last one is
the same as `--commits=<N>` flag that some commands supported previously.

The default is `--scope=diff_to_main`, which finds all the files that were
modified in this branch compared to the main branch, and uses that to find all
the crates that are affected by these modifications. The commands are only run
for the affected crates.

Updates the `.github/workflows/ci.yaml` to fetch the last two commits in the PR.
This allows computing the diff, and running the cargo-related commands only for
the necessary crates. For `pull_request` events, the CI steps are run with the
`--scope=commits:1` flag and for the `push` events, they are run with the
`--scope=all` flag.