Add UwabiExtension Trait and Testing extension (#2501)
Add enum to `BoxedExtension` to distinguish between Uwabi and Native extension. An extension factory can currently produce either one of the two (but fixed which one, which allows minimal changes to existing extensions). Add Uwabi Extension Trait and implement it with a Testing extension which requires adding a Testing Channel Handle. Add Uwabi extension endpoints to WasmState (requires to use ChannelHandles as keys, because Uwabi Extension do not have a abi_function_name).