Define relation between OakStatus and ChannelStatus (#2512)
Rename and refactor map_host_errors to allow to get a result from a ChannelStatus as well as an OakStatus. Change return type to ChannelStatus for Wasm memory manipulation, because they are more related to Channels than to Oak, which required to inverse From relationship between OakStatus and ChannelStatus. Add channel_read/write to invoke index.