Use a custom implementation of Mutex instead of std (Fixes #2579)
This is in preparation for the Oak Functions loader compatible with no_std
in the future.

Note that the current use of RwLock was replaced by a Mutex. This is because
the critical section is extremely short: just cloning or replacing an Arc. Such
short critical sections do not need the flexibility of reader-writer lock.