Remove "Developing With OS X" docs
It contains outdated references to task ("run-examples", has been renamed in #2612). Instead of updating this, I think we can remove this doc? I don't think it makes sense to maintain this any longer. None of us use OSX and for OSX users it's probably easier to just use the dev container