Use vendored version of ring in remote attestation  (#2661)
* Clarify cargo license for all ring versions

* Patch the ring build script to enable builds from source code

Turns out that ring includes pre-generated objects in cargo releases. It uses logic in its build scripts to determine whether the build is from a version published on cargo, or from source code. This logic is somewhat naively implemented by checking for the absence of a `.git` directory. This check falesly leads ring to believe it is building from a version published on crates.io, and fails when it tries to include the pregenerated files not present in the source code. To build from source code we must patch this flag.

* Use the vendored version of ring

Requires updating our code to accomodate API changes since the vendored version is newer

* bust CI cache

* Allow deprecated code inside of vendored ring. It includes code that is, and our CI fails on warnings