Move request logic into a separate crate (#2745)
This moves the trivial echo code from the UEFI app code into a separate crate.

The goals are two-fold:
(a) First, the UEFI app crate should only be kind of hardware abstraction layer, with no cleverness of its own. All logic should be somewhere else.
(b) Second, this allows us to create a bare-metal 'kernel' which offers a similar abstraction layer, and use the common runtime library shared with the UEFI app.