Add an end-to-end test for bare metal/UEFI using xtask (#2803)
This comes with multiple caveats:

KVM is required, so you can't run it in the VS Code container nor any of our current CI environments
Bare-metal binary doesn't build, so that branch is untested (but as we know it'd fail anyway due to ring)
Next up: setting up a Kokoro run for this test.

In the long term I'd assume this test will go away in its entirety and will be subsumed by the run-oak-function-examples test. But for now it's better than nothing.