Oak IDL based on flatbuffers
The goal of the IDL is to be used both as part of the communication
channel between trusted and untrusted runtime, as well as for the
communication between Wasm modules and the trusted runtime.

Create the following new crates:

- `oak_idl_gen_services`
- `oak_idl_gen_structs`
- `oak_idl_gen_tests`

Align the existing `oak_idl` crate to use flatbuffers via the
aforementioned crates.

Import flatbuffers reflection schema in `third_party`.

Remove unused deps throughout to make cargo udeps and cargo deny happy.

Ref #2500