Use gRPC status codes in the oak IDL
Tiziano proposed this earlier, and I agree it's probably a good idea. Crucially gRPC Status codes include a definition of an Ok code. This is nice, as it means we can use status codes as is in the framing implementation.

The downside is that technically an invocation could error with an Ok status, which wouldn't make sense. That's the same type / error signature as used by tonic though, so it's probably fine