Include all communication channels in one binary (#3014)
We support three communication channels -- virtio-console, virtio-vsock, and serial -- but right now choosing between them is with mutually incompatible features. That's not how features are supposed to work, for multiple reasons: features are supposed to be additive (otherwise --all-features turns into nonsense), and testing is difficult, and it's too easy to accidentally break code.

Now that we've added a facility to handle kernel command line arguments, let's make use of it to choose between the three comms channels, and build support for all three into the kernel binary.

  - You can still enable and disable them with features as you wish. In fact, you can disable all three, if you want to.
  - If you enable only one channel type, it will be used by default without the need to pass in command line args.
  - If you enable multiple channel types, you can switch between them with a command line arg.
  - The features are additive, so --all-features does the correct thing.

Unfortunately this means that we won't know the size of the struct that implements said channel statically during build, so we have to Box<> things and carry it around everywhere.