Use rootless Docker
Required disabling buildkit support, which seems incompatible with
rootless for some reason (needs further investigation in #3225).

The next step (if no one complains) will be to remove the
`fix_docker_user_and_run` script, in another PR.

This should not affect CI, since CI usually already runs under root
anyways (Docker is not a security boundary, so it would be unlikely that
any CI would run the Docker daemon as root but the CI steps as a
different user).

Fix #2602