Remove unsafe version of Oak Functions (#3245)
Move all tests to use the base version.

Enable support for grpc-web in the base version too; this will not be
around long there anyways, as we transition to using the split launcher
anyways, which will implement it on the untrusted side (therefore will
not affect the trusted runtime at all).

Fix #3242