Use Split launcher to run examples
Change logic in `xtask` to parse the config values (e.g. wasm module
location, constant size) from the example.toml file directly, since the
split launcher accepts them as separate flags instead of a single config
file.

The web client does not seem to work with the split launcher, so I
temporarily disabled it and created #3256 to follow up.

The Java client was also initially not working, but I tracked that down
to the server trying to parse the transcript signing key, but the Java
client only providing zeros for it at the moment. That work is already
tracked in #2918.

Change provenance generation to use the crosvm version of the trusted
runtime, which is the one we are ultimately interested in.

Fix #3241