Create new fuzz target to replace existing one
The oak_functions_loader crate will be refactored / removed as part of
 #3082, so this maintains a fuzz target with similar semantics, in a
 location where it is expected to stay untouched. The new fuzz logic is
 somewhat simpler, but it boils down to testing the same policy,
 although more directly.

This PR creates a new crate as the `fuzz` top-level directory,
so that it will be easier in the future to add new fuzz targets directly
there without having to synchronize the oss-fuzz configuration.

The previous fuzz target will be removed once the oss-fuzz configuration
is updated to point to the newly created one.

Ref #3082