Use split Oak Functions server for tests
The overall goal is to stop depending on the unified Oak Functions
loader library, so that it may be deleted.

Also remove some of the tests that rely on constant processing time,
since that's not supported in the split binary yet (#3079).

Incidentally, I made the integration tests more aligned with actually
testing meaningful integrations, and I am starting to think that they
can replace running the examples at some point in the future (so we
drive the examples from proper Rust tests, instead of doing that from xtask).

Ref #3081