Implement IDL based on protobuf
Currently as a stand-alone crate, but in the future it may replace (and
be renamed into) `oak_idl`.

Currently not used anywhere, just in its own tests.

The structure of the code generator is identical to the existing one
from
https://github.com/project-oak/oak/blob/e75740a22c5c3db3d25031cd8b660b50aa60eda2/oak_idl_gen_services/src/lib.rs
, which I copy-pasted and modified in order to be driven from prost
instead of flatc.

The easiest way of reviewing this PR is probably to look at the
generated file, which is also checked in for convenience and to catch
regression or changes in the generator.

Ref #3385