Serialize requests and responses via IDL
This allows serializing the combination of method id and request body
into an IDL message (so meta), which is then sent over the underlying
transport, which now only needs to know about raw bytes and nothing
about the serialization mechanism.

As a response, also the errors from the service are serialized back via
the IDL to the client, so that we can have something much more similar
to how RPC / IPC mechanisms usually work.

This adds an additional layer of encoding / decoding, but since it's the
same as the messages in the inner layer, it seems acceptable after all.

Eventually the oak_protobuf_idl will replace the oak_idl crate, so the
duplication will go away.

Ref #3385