Set the default caching mode to write-protect (WP).
The Linux kernel, since July of 2022, crashes if MTRR is not set to WP, in a check called only if some form of SEV memory encryption is enabled.  This change has been tested to not impact lanching the 5.15.103 kernel when not in SEV on an Intel processor, and also that the kernel boots properly with SEV enabled on AMD Milan processors.