Add IREE C++ library template (#3734)
This PR adds an IREE (https://github.com/openxla/iree) service for Oak.

Currently it links together a pre-compiled IREE library and a Rust service.

Because the library is currently compiled internally without using Oak Toolchain, we have to link `newlib` manually and put compiled `*.a` files in the `cc/iree` directory.

Also because we link Rust and C++ together, we need to make the Restricted Kernel use a different location for heap, to not overlap with the C++ heap. Currently it uses `0x100_0000_0000` as a Rust heap location. This limitation will be removed once we support building C++ applications internally using the Oak Toolchain (https://github.com/project-oak/oak/issues/3864).