A bash script that prepares minirootfs + docker image as initramfs. (#3896)
This commit has a a bunch of changes:

1. A `prepare_docker_launcher_initramfs.sh` script combines the following to create the initramfs:
   - Alpine's minirootfs
   - Necessary drivers from an Alpine Linux distribution.
   - Filesystem extracted from the given docker image.
Note that if the docker image is not provided, the script creates an initramfs that provides a working shell.
1. Generalize the `docker_to_initramfs.sh` script to generate a chroot directory or initramfs images
1. A hack to set up PATH environment variables to some common locations.

Note that this PR is meant to demonstrate the steps that are needed to launch a docker container with networking support. To achieve the task, we pull drivers and kernel from the [Alpine Linux](www.alpinelinux.org) distribution. We also use Alpine's minirootfs as the docker launcher. Ideally, we will have to build the Linux kernel, drivers, and a docker launcher from scratch. However, this prototype provides a good step towards making progress on the Oak Docker project. For example, using this prototype, we were able to launch and interact with a docker container running Tensorflow Jupyter notebook running TF+CPU runtime.