Add JNI code for hpke crypto (#3996)
This PR adds the JNI code for HPKE seal and open methods to call the implementations in c++ from java.

The implementation of the JNI functions is pending the C++ implementation of related classes to call and currently just mimics the existing java implementation. The JNI code for HPKE currently creates empty Context objects in C++ and returns the pointer to Java. The C++ pointer is then passed in the seal/open JNI calls.

This PR is part of the the Java implementation for (#3642).

Added TODOs to code where pending the C++ implementation.