Oak Containers Hello World: Expose service in Trusted App
This inverts the prior hello world example in which the trusted app implemented a gRPC client that invoked a gRPC service exposed by the untrusted app.

Now the trusted app exposes the service and the untrusted app invokes it, which is closer to a real world example of Oak Containers.