Refactor Oak Functions to separate encryption (#4199)
Refactor the core logic of Oak Functions into a separate `OakFunctionsInstance` object. This object is still in the same crate as `OakFunctionsService` for now, but does not have any logic related to remote attestation nor encryption at all. Instead, that logic now only exists in the `OakFunctionsService` wrapper.

I changed the `InitializationState` from an enum to an `Option` so that it behaves like a proper functor (being able to use map and other methods on it).

`AttestationHandler` is changed to take an `FnOnce` and is expected to be constructed on the fly around a single invocation of an individual method, therefore it does not need to own the entire object being invoked, but can borrow a specific method when needed (and could be applied to several methods if necessary). The way around the ownership in the previous version was to have an `Arc<LookupDataManager>` instance shared between the outer service and the initialized instance, which is not a generally applicable pattern, while the current pattern should be fully generic.

The methods that do not require encryption defer to the inner methods directly (there may be a way of reducing the boilerplate in the future), while `invoke` creates a temporary `AttestationSessionHandler` object and uses it immediately for the invocation, then discards it.

Reorder initialization of Oak Functions so that lookup data is only sent after it is actually initialized.

Ref #4182