Add initial bzImage-compatible wrapper (#4260)
This is a very simple wrapper for the Oak Restricted Kernel that looks similar enough the the Linux bzImage format so that it can be loaded by QEMU as if it was a Linux kernel. It is compatible with the 64-bit Linux boot protocol, so can be booted by Stage 0.

This initial version has many limitations. See #4319.

These limitations will be addressed in follow-up PRs.