Rename RestrictedKernelDiceData for clarity (#4383)
* Rename RestrictedKernelDiceData for clarity

The previous name RestrictedKernelDiceData was inaccurate, since in our naming scheme dice data denotes a struct that in addition to evidence, includes a certificate authoruty that may be used to construct another dice layer. This is not the case for the application, which in our usecase represents the final dice layer. It should not be allowed to create a new layer on top of itself. The keys it is provisioned with are for signing and encryption only.

* address comments

* fixups