Make FD read behaves consistently & idiomatically
Now all FDs keep an inner state that tracks how much they've been read. Subsequent reads return the subequent bytes. If the FD has been read to end, no bytes are copied on subsequent reads.

If we want to permit mulitple reads in the future, we should add a seek method to permit the app manipulate the state of the FD.