Make all FD reads behave consistently & idiomatically (#4393)
* zero out DiceData after reading it and keep state

addresses https://github.com/project-oak/oak/pull/4377#discussion_r1361345702

I'm unsure if this is the right way.

Pro: I think zeroing out the private key is a good idea, and I think it should be done automatically once it's read.

Con: The stateful behavior of this descriptor is different from all other ones. They're stateless and always read from 0

* fix logic

* fix minor logic error

* Make FD read behaves consistently & idiomatically

Now all FDs keep an inner state that tracks how much they've been read. Subsequent reads return the subequent bytes. If the FD has been read to end, no bytes are copied on subsequent reads.

If we want to permit mulitple reads in the future, we should add a seek method to permit the app manipulate the state of the FD.

* remove redundant var

* fix failing test