Impl signing crypto oracle in the restricted kernel sdk
Previously we would expose the signign key directly, now it's behind a struct whose only public method is a sign.

Tracking http://b/314749025 and http://b/312897453.

The implementation is slightly different from what is outlined in the bugs. Specifically the bug propsed implementing a create method directly Signer struct, which would call directly into the kernel to fetch the key.

This isn't feasible dice data can only be read from the restricted kernel once and only read at once. Hence a create method is used for the overarching DiceWrapper struct.

If we want to maintain the orginally proposed interface, we could store the dice data seperately in memory for the application, and read ot from there whenever a new struct is created. A perhaps more ergonimic but also more complex and magical solution. Hence why I didn't go with it. Thoughts otherwise are welcome in the comments