Clean up oci_runtime_bundle temporary files. (#4695)
* Clean up oci_runtime_bundle temporary files.

bazel doesn't handle this automatically. As-is, the right outputs are produced and there aren't any warnings or errors, but temporary files do accumulate in /tmp. They'll go away on reboot, but that may not happen that frequently on all machines. Cleaning up after ourselves is straightforward and a best practice.