Extract restricted kernel specific logic from oak_functions_service
This extracts logic specific to the restricted kernel from the `oak_functions_service` crate, similiarly to what we do for oak container's in oak_functions_containers_app/src/lib.rs.

It fixes an issue encountered in the oak container's version, which depends on a submodule of the `oak_functions_service` crate, and inherited it's dependency on the `oak_restricted_kernel_sdk.