SDK, reorganize mod structure for clarity
Moves a couple of utility functions from the root level into a pub channel and utils mod. Should make the docs a lot easier to read. Also aligns the crate file structure with its pub interface for clarity.

This is technically a breaking change, but all moved functions are very new, and I don't expect teams to already be depending on them. And if they are, the compiler error message should make migration not too difficult.