Add initrd feature flag to kernel
At this point we imo need conditional compilation, because we want the kernel to behave quite differently based on whether it is loading an applicatino over the channel vs from the initrd.

In the first case we want to create a new dice layer. In the second case we want to just expose the stage0 dice data to the orchestrator